+++ odbc/odbc_handles.h
@@ -207,6 +207,9 @@ public:
                        const std::optional<std::string>& schema_name,
                        const std::string& table_name,
                        bool unique_only);
+  SQLRETURN procedures(const std::optional<std::string>& catalog_name,
+                       const std::optional<std::string>& schema_name,
+                       const std::optional<std::string>& procedure_name);
   SQLRETURN tables(const std::optional<std::string>& catalog_name,
                    const std::optional<std::string>& schema_name,
                    const std::optional<std::string>& table_name,